django_backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
python-dotenv==1.0.0
openpyxl==3.1.2
google-api-python-client==2.111.0
google-auth==2.26.1
Pillow==10.2.0
